pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-ollama>=0.3.10",
    "pypdf>=6.1.1",
    "python-dotenv>=1.0.0",
]
